# src/main/resources/latest-spark-swagger-output.yaml
...
swagger: "2.0"
info:
  description: "OpenTripPlanner Middleware API"
  title: "OTP Middleware"
  termsOfService: ""
  contact:
    name: "IBI Group"
    url: "https://github.com/ibi-group"
    email: "transitdatasupport@ibigroup.com"
  license:
    name: "MIT License"
    url: "https://opensource.org/licenses/MIT"
host: "localhost:4567"
basePath: "/"
...
schemes:
- "https"
paths:
...
  /api/admin/user/{id}:
    get:
      tags:
      - "api/admin/user"
      description: "Returns the 'AdminUser' entity with the specified id, or 404 if\
        \ not found."
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The id of the entity to search."
        required: true
        type: "string"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/AdminUser"
          responseSchema:
            $ref: "#/definitions/AdminUser"
    put:
      tags:
      - "api/admin/user"
      description: "Updates and returns the 'AdminUser' entity with the specified\
        \ id, or 404 if not found."
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The id of the entity to update."
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "Body object description"
        required: true
        schema:
          $ref: "#/definitions/AdminUser"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/AdminUser"
          responseSchema:
            $ref: "#/definitions/AdminUser"
    delete:
      tags:
      - "api/admin/user"
      description: "Deletes the 'AdminUser' entity with the specified id if it exists."
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The id of the entity to delete."
        required: true
        type: "string"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/AdminUser"
          responseSchema:
            $ref: "#/definitions/AdminUser"
...
  /api/secure/application/authenticate:
    post:
      tags:
      - "api/secure/application"
      description: "Authenticates ApiUser with Auth0."
      produces:
      - "application/json"
      parameters:
      - name: "username"
        in: "query"
        description: "Auth0 username (usually email address)."
        required: true
        type: "string"
      - name: "password"
        in: "query"
        description: "Auth0 password."
        required: true
        type: "string"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/TokenHolder"
          responseSchema:
            $ref: "#/definitions/TokenHolder"
...
  /api/secure/application:
    get:
      tags:
      - "api/secure/application"
      description: "Gets a paginated list of all 'ApiUser' entities."
      produces:
      - "application/json"
      parameters:
      - name: "limit"
        in: "query"
        description: "If specified, the maximum number of items to return."
        required: false
        type: "string"
        default: "10"
      - name: "offset"
        in: "query"
        description: "If specified, the number of records to skip/offset."
        required: false
        type: "string"
        default: "0"
      - name: "userId"
        in: "query"
        description: "If specified, the required user id."
        required: false
        type: "string"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/ResponseList"
          responseSchema:
            $ref: "#/definitions/ResponseList"
    post:
      tags:
      - "api/secure/application"
      description: "Creates a 'ApiUser' entity."
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Body object description"
        required: true
        schema:
          $ref: "#/definitions/ApiUser"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiUser"
          responseSchema:
            $ref: "#/definitions/ApiUser"
  /api/secure/application/{id}:
    get:
      tags:
      - "api/secure/application"
      description: "Returns the 'ApiUser' entity with the specified id, or 404 if\
        \ not found."
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The id of the entity to search."
        required: true
        type: "string"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiUser"
          responseSchema:
            $ref: "#/definitions/ApiUser"
    put:
      tags:
      - "api/secure/application"
      description: "Updates and returns the 'ApiUser' entity with the specified id,\
        \ or 404 if not found."
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The id of the entity to update."
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "Body object description"
        required: true
        schema:
          $ref: "#/definitions/ApiUser"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiUser"
          responseSchema:
            $ref: "#/definitions/ApiUser"
    delete:
      tags:
      - "api/secure/application"
      description: "Deletes the 'ApiUser' entity with the specified id if it exists."
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The id of the entity to delete."
        required: true
        type: "string"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiUser"
          responseSchema:
            $ref: "#/definitions/ApiUser"
...
  /api/secure/user/{id}:
    get:
      tags:
      - "api/secure/user"
      description: "Returns the 'OtpUser' entity with the specified id, or 404 if\
        \ not found."
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The id of the entity to search."
        required: true
        type: "string"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/OtpUser"
          responseSchema:
            $ref: "#/definitions/OtpUser"
    put:
      tags:
      - "api/secure/user"
      description: "Updates and returns the 'OtpUser' entity with the specified id,\
        \ or 404 if not found."
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The id of the entity to update."
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "Body object description"
        required: true
        schema:
          $ref: "#/definitions/OtpUser"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/OtpUser"
          responseSchema:
            $ref: "#/definitions/OtpUser"
    delete:
      tags:
      - "api/secure/user"
      description: "Deletes the 'OtpUser' entity with the specified id if it exists."
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The id of the entity to delete."
        required: true
        type: "string"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/OtpUser"
          responseSchema:
            $ref: "#/definitions/OtpUser"
...
definitions:
  AdminUser:
    type: "object"
    properties:
      subscriptions:
        type: "array"
        items:
          type: "string"
          enum:
          - "NEW_ERROR"
...
  ResponseList:
    type: "object"
    properties:
      clazz:
        type: "string"
      data:
        type: "array"
        items:
          $ref: "#/definitions/Object"
      offset:
        type: "integer"
        format: "int32"
      limit:
        type: "integer"
        format: "int32"
      total:
        type: "integer"
        format: "int64"
      timestamp:
        type: "string"
        format: "date"
  ApiKey:
    type: "object"
    properties:
      keyId:
        type: "string"
      name:
        type: "string"
      value:
        type: "string"
  ApiUser:
    type: "object"
    properties:
      apiKeys:
        type: "array"
        items:
          $ref: "#/definitions/ApiKey"
      appName:
        type: "string"
      appPurpose:
        type: "string"
      appUrl:
        type: "string"
      company:
        type: "string"
      hasConsentedToTerms:
        type: "boolean"
      name:
        type: "string"
  TokenHolder:
    type: "object"
    properties:
      accessToken:
        type: "string"
      idToken:
        type: "string"
      refreshToken:
        type: "string"
      tokenType:
        type: "string"
      expiresIn:
        type: "integer"
        format: "int64"
...
  UserLocation:
    type: "object"
    properties:
      address:
        type: "string"
      icon:
        type: "string"
      lat:
        type: "number"
        format: "double"
      lon:
        type: "number"
        format: "double"
      name:
        type: "string"
      type:
        type: "string"
  OtpUser:
    type: "object"
    properties:
      accessibilityRoutingByDefault:
        type: "boolean"
      hasConsentedToTerms:
        type: "boolean"
      isPhoneNumberVerified:
        type: "boolean"
      notificationChannel:
        type: "string"
      phoneNumber:
        type: "string"
      savedLocations:
        type: "array"
        items:
          $ref: "#/definitions/UserLocation"
      storeTripHistory:
        type: "boolean"
      applicationId:
        type: "string"
...